Move some common functions to lib
