+++ pupy/modules/ls.py
@@ -3,6 +3,7 @@ from pupylib.PupyModule import *
 from pupylib.PupyCompleter import remote_path_completer
 from pupylib.utils.rpyc_utils import obtain
 from pupylib.utils.term import colorize
+from modules.lib import size_human_readable, file_timestamp, to_utf8
 from datetime import datetime
 
 __class_name__="ls"
@@ -19,32 +20,6 @@ T_PATH      = 8
 T_FILES     = 9
 T_FILE      = 10
 T_TRUNCATED = 11
-
-def file_timestamp(timestamp):
-    try:
-        d = datetime.fromtimestamp(timestamp)
-        return str(d.strftime("%d/%m/%y"))
-    except:
-        return '00/00/00'
-
-def size_human_readable(num, suffix='B'):
-    try:
-        num = int(num)
-        for unit in ['','Ki','Mi','Gi','Ti','Pi','Ei','Zi']:
-            if abs(num) < 1024.0:
-                return "%3.1f %s%s" % (num, unit, suffix)
-            num /= 1024.0
-        return "%.1f %s%s" % (num, 'Yi', suffix)
-    except:
-        return '0.00 B'
-
-def to_utf8(value):
-    if type(value) == unicode:
-        return value
-    try:
-        return value.decode('utf-8')
-    except:
-        return value.decode('latin1', errors='ignore')
 
 def output_format(file, windows=False):
     if windows:
